requirements.txt
streamlit
google-generativeai
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
faiss-cpu
pypdf
sqlalchemy
psycopg2-binary
python-dotenv